# specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Semiont API",
    "version": "0.1.0",
    "description": "Semantic Knowledge Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "schemas": {
      "ContentFormat": {
        "type": "string",
        "enum": [
          "text/plain",
          "text/markdown"
        ],
        "description": "Content format as MIME type"
      },
      "Motivation": {
        "type": "string",
        "enum": [
          "assessing",
          "bookmarking",
          "classifying",
          "commenting",
          "describing",
          "editing",
          "highlighting",
          "identifying",
          "linking",
          "moderating",
          "questioning",
          "replying",
          "tagging"
        ],
        "description": "W3C Web Annotation motivation vocabulary - https://www.w3.org/TR/annotation-vocab/#motivation"
      },
...
      "EventStreamResponse": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "event",
          "data"
        ]
      },
...
      "GenerateResourceRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "Resource ID containing the annotation"
          },
          "title": {
            "type": "string",
            "description": "Custom title for generated resource"
          },
          "prompt": {
            "type": "string",
            "description": "Custom prompt for content generation"
          },
          "language": {
            "type": "string",
            "description": "Language locale (e.g., \"es\", \"fr\", \"ja\")"
          }
        },
        "required": [
          "resourceId"
        ]
      },
...
      "GetAnnotationResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedResource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "annotation",
          "resource",
          "resolvedResource"
        ]
      },
...
      "GetAnnotationHistoryResponse": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "timestamp": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                },
                "resourceId": {
                  "type": "string"
                },
                "payload": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "format": {
                          "$ref": "#/components/schemas/ContentFormat"
                        },
                        "contentChecksum": {
                          "type": "string"
                        },
                        "creationMethod": {
                          "type": "string",
                          "enum": [
                            "api",
                            "upload",
                            "ui",
                            "reference",
                            "clone",
                            "generated"
                          ]
                        },
                        "entityTypes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "metadata": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "name",
                        "format",
                        "contentChecksum",
                        "creationMethod"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "format": {
                          "$ref": "#/components/schemas/ContentFormat"
                        },
                        "contentChecksum": {
                          "type": "string"
                        },
                        "parentResourceId": {
                          "type": "string"
                        },
                        "creationMethod": {
                          "type": "string",
                          "enum": [
                            "api",
                            "upload",
                            "ui",
                            "reference",
                            "clone",
                            "generated"
                          ]
                        },
                        "entityTypes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "metadata": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "name",
                        "format",
                        "contentChecksum",
                        "parentResourceId",
                        "creationMethod"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "reason": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {}
                    },
                    {
                      "type": "object",
                      "properties": {
                        "annotation": {
                          "$ref": "#/components/schemas/Annotation",
                          "description": "W3C Annotation (creator/created fields are optional in events - they're derived from event userId/timestamp)"
                        }
                      },
                      "required": [
                        "annotation"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "annotationId": {
                          "type": "string",
                          "description": "Unified annotation ID"
                        }
                      },
                      "required": [
                        "annotationId"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "annotationId": {
                          "type": "string",
                          "description": "Unified annotation ID"
                        },
                        "targetResourceId": {
                          "type": "string",
                          "description": "For linking motivation"
                        }
                      },
                      "required": [
                        "annotationId",
                        "targetResourceId"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "entityType": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "entityType"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "entityType": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "entityType"
                      ]
                    }
                  ]
                },
                "metadata": {
                  "type": "object",
                  "properties": {
                    "sequenceNumber": {
                      "type": "number"
                    },
                    "prevEventHash": {
                      "type": "string"
                    },
                    "checksum": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "sequenceNumber"
                  ]
                }
              },
              "required": [
                "id",
                "type",
                "timestamp",
                "userId",
                "resourceId",
                "payload",
                "metadata"
              ]
            }
          },
          "total": {
            "type": "number"
          },
          "annotationId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          }
        },
        "required": [
          "events",
          "total",
          "annotationId",
          "resourceId"
        ]
      },
...
      "DeleteAnnotationRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "format": "uri",
            "description": "Resource URI containing the annotation (required for O(1) Layer 3 lookup)"
          }
        },
        "required": [
          "resourceId"
        ]
      },
...
      "Annotation": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "string",
            "const": "http://www.w3.org/ns/anno.jsonld",
            "default": "http://www.w3.org/ns/anno.jsonld",
            "description": "W3C Web Annotation JSON-LD context"
          },
          "type": {
            "type": "string",
            "const": "Annotation",
            "default": "Annotation",
            "description": "W3C Annotation type"
          },
          "id": {
            "type": "string"
          },
          "motivation": {
            "$ref": "#/components/schemas/Motivation"
          },
          "target": {
            "oneOf": [
              {
                "type": "string",
                "description": "IRI of the entire resource being annotated"
              },
              {
                "$ref": "#/components/schemas/AnnotationTarget"
              }
            ],
            "description": "W3C Web Annotation target - can be a simple IRI string (entire resource) or an object with source and optional selector (fragment)"
          },
          "body": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "maxItems": 0,
                "description": "Empty array for stub references (Phase 1 pattern)"
              },
              {
                "$ref": "#/components/schemas/AnnotationBody",
                "description": "Single body (TextualBody or SpecificResource)"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "description": "Array of mixed TextualBody (tagging) and SpecificResource (linking) bodies"
              }
            ],
            "description": "W3C Web Annotation body - can be empty array, single body, or array of bodies"
          },
          "creator": {
            "$ref": "#/components/schemas/Agent"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          },
          "generator": {
            "$ref": "#/components/schemas/Agent"
          }
        },
        "required": [
          "@context",
          "type",
          "id",
          "motivation",
          "target",
          "body"
        ]
      },
      "AnnotationBody": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/TextualBody"
          },
          {
            "$ref": "#/components/schemas/SpecificResource"
          }
        ],
        "description": "Phase 2: Body can be TextualBody (for entity tags, descriptions) or SpecificResource (for links)"
      },
...
    },
...
  },
  "paths": {
...
    "/annotations/{id}": {
      "get": {
        "summary": "Get Annotation by URI (W3C Content Negotiation)",
        "description": "W3C-compliant globally resolvable annotation URI with content negotiation. Returns JSON-LD for API clients (Accept: application/ld+json or application/json). Redirects to frontend for browsers (Accept: text/html or browser User-Agent). Requires authentication and resourceId query parameter.",
        "tags": [
          "W3C URIs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Annotation ID"
          },
          {
            "name": "resourceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Resource ID or URI containing the annotation"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation representation (JSON-LD)",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnnotationResponse"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to frontend (for browsers)"
          },
          "400": {
            "description": "Missing resourceId parameter"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
    "/api/annotations/{id}/generate-resource-stream": {
      "post": {
        "summary": "Generate Resource from Reference (SSE)",
        "description": "Stream real-time resource generation progress via Server-Sent Events",
        "tags": [
          "Selections",
          "Resources",
          "Real-time",
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Reference/annotation ID"
            },
            "required": true,
            "description": "Reference/annotation ID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream opened successfully",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStreamResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Reference not found"
          }
        }
      }
    },
...
    "/api/annotations/{id}": {
      "delete": {
        "summary": "Delete Annotation",
        "description": "Delete an annotation (requires resourceId in body for O(1) Layer 3 lookup)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Annotation deleted successfully"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
    "/api/resources/{resourceId}/annotations/{annotationId}/history": {
      "get": {
        "summary": "Get Annotation History",
        "description": "Get full event history for a specific annotation (highlight or reference)",
        "tags": [
          "Selections",
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "resourceId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "annotationId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnnotationHistoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
  },
...
}